pyproject.toml
[tool.poetry]
name = "highroler"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
"discord.py" = "^1.7.3"
ZODB = "^5.7.0"
python-dotenv = "^0.20.0"

[tool.poetry.dev-dependencies]

[tool.poetry.scripts]
bot = "highroler:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
